requirements.txt
streamlit
python-dotenv
pytesseract
pdf2image
Pillow
transformers
torch
